LPS-79878 Moving verifyArticleAssets to a single upgrade step since the upgrade framework works better with small steps. It has been moved to the step 0.0.8 -> 1.0.0 since this verification was added by LPS-51384 just before releasing DXP
